platform/pyproject.toml
[project]
name = "platform-app"
version = "1.4.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "bcrypt>=4.1",
    "boto3>=1.35",
    "python-multipart>=0.0.9",
    "psycopg[binary]>=3.2",
    "redis>=5.0",
    "duckdb>=1.1",
    "pyyaml>=6.0",
    "pandas>=2.2",
    "openpyxl>=3.1",
    "pyarrow>=15.0",
    "croniter>=2.0",
    "markdown-it-py>=3.0",
    "anthropic>=0.92",
    "pdfplumber>=0.11",
    # yunwei_win (智通客户) deps — vendored from yunwei-tools.
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "aiosqlite>=0.20",
    "pydantic-settings>=2.6",
    "pypdf>=5.1",
    "tenacity>=9.0",
    "landingai-ade>=0.1.0",
    "rq>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
]

[project.scripts]
platform-admin = "platform_app.admin:main"
# Background worker for /win/api/ingest/jobs. Run as a separate Railway
# service with `Start Command: yunwei-win-ingest-worker` and the same env as
# the web service (DATABASE_URL, REDIS_URL, DATA_ROOT, MISTRAL_API_KEY,
# VISION_AGENT_API_KEY, DOCUMENT_AI_PROVIDER, ANTHROPIC_API_KEY, …).
yunwei-win-ingest-worker = "yunwei_win.workers.ingest_rq_worker:main"

[tool.pytest.ini_options]
asyncio_mode = "strict"
# Add the worktree's platform/ directory to sys.path so tests import
# platform_app from the local working tree rather than the editable install
# pointing at the main worktree (which uses a shared .venv).
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Ship both top-level packages: platform_app (orchestration) and yunwei_win
# (vendored yunwei-tools backend powering /win/api/*).
packages = ["platform_app", "yunwei_win"]
